Build an actual layout tree.

+++ src/html_parser.rs
@@ -205,7 +205,7 @@ fn convert_html_nodes_to_dom(html_node: HtmlNode) -> Document {
 
     //TODO: figure out a good way to assert things, with no runtime (release) costs, and possible to disable in debug mode as well... for now its always there
     //      we see to have something already in debug.rs
-    if document_dom_nodes.len() != next_node_internal_id as usize {
+    if document_dom_nodes.len() != next_node_internal_id {
         panic!("Id seting of DOM nodes went wrong");
     }
